+++ j_backend/src/main/java/com/antock/api/member/application/service/MemberApplicationService.java
@@ -10,9 +10,9 @@ import com.antock.api.member.value.MemberStatus;
 import com.antock.api.member.value.Role;
 import com.antock.global.common.exception.BusinessException;
 import com.antock.global.common.exception.ErrorCode;
-import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -20,7 +20,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-@RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class MemberApplicationService {
 
@@ -29,64 +28,61 @@ public class MemberApplicationService {
     private final MemberDomainService memberDomainService;
     private final AuthTokenService authTokenService;
     private final RateLimitServiceInterface rateLimitService;
-    private final MemberCacheService memberCacheService;
+    private final MemberCacheService memberCacheService; // 인터페이스 사용
     private final PasswordEncoder passwordEncoder;
+
+    @Autowired
+    public MemberApplicationService(MemberDomainService memberDomainService,
+                                    AuthTokenService authTokenService,
+                                    RateLimitServiceInterface rateLimitService,
+                                    @Autowired(required = false) MemberCacheService memberCacheService,
+                                    PasswordEncoder passwordEncoder) {
+        this.memberDomainService = memberDomainService;
+        this.authTokenService = authTokenService;
+        this.rateLimitService = rateLimitService;
+        this.memberCacheService = memberCacheService;
+        this.passwordEncoder = passwordEncoder;
+    }
 
     @Transactional
     public MemberResponse join(MemberJoinRequest request) {
-        // Rate Limiting 체크
         rateLimitService.checkRateLimit(request.getUsername(), "join");
-
         String encodedPassword = passwordEncoder.encode(request.getPassword());
-
         Member member = memberDomainService.createMember(
                 request.getUsername(),
                 encodedPassword,
                 request.getNickname(),
                 request.getEmail()
         );
-
         MemberResponse response = MemberResponse.from(member);
-
-        memberCacheService.cacheMemberResponse(response);
-
-        log.info("회원 가입 완료 - username: {}, id: {}",
-                member.getUsername(), member.getId());
-
+        if (memberCacheService != null) {
+            memberCacheService.cacheMemberResponse(response);
+        }
+        log.info("회원 가입 완료 - username: {}, id: {}", member.getUsername(), member.getId());
         return response;
     }
 
     @Transactional
     public MemberLoginResponse login(MemberLoginRequest request, String clientIp) {
-
         rateLimitService.checkRateLimit(clientIp, "login");
-
         Member member = memberDomainService.findByUsername(request.getUsername())
                 .orElseThrow(() -> new BusinessException(ErrorCode.INVALID_CREDENTIALS));
-
-
         validateMemberStatus(member);
-
-
         if (!passwordEncoder.matches(request.getPassword(), member.getPassword())) {
             memberDomainService.handleLoginFailure(member);
-
-            memberCacheService.evictMemberCache(member.getId());
-
+            if (memberCacheService != null) {
+                memberCacheService.evictMemberCache(member.getId());
+            }
             throw new BusinessException(ErrorCode.INVALID_CREDENTIALS);
         }
-
         memberDomainService.handleLoginSuccess(member);
-
         MemberResponse memberResponse = MemberResponse.from(member);
-        memberCacheService.cacheMemberResponse(memberResponse);
-
+        if (memberCacheService != null) {
+            memberCacheService.cacheMemberResponse(memberResponse);
+        }
         String accessToken = authTokenService.generateAccessToken(member);
         String refreshToken = authTokenService.generateRefreshToken(member);
-
-        log.info("로그인 성공 - username: {}, id: {}",
-                member.getUsername(), member.getId());
-
+        log.info("로그인 성공 - username: {}, id: {}", member.getUsername(), member.getId());
         return MemberLoginResponse.builder()
                 .member(memberResponse)
                 .accessToken(accessToken)
@@ -95,67 +91,60 @@ public class MemberApplicationService {
                 .build();
     }
 
-
     public MemberResponse getCurrentMemberInfo(Long memberId) {
         log.debug("현재 사용자 정보 조회 요청 - ID: {}", memberId);
-
-        MemberResponse cachedMember = memberCacheService.getMemberFromCache(memberId);
-        if (cachedMember != null) {
-            log.debug("캐시에서 현재 사용자 정보 조회 성공 - ID: {}", memberId);
-            return cachedMember;
+        if (memberCacheService != null) {
+            MemberResponse cachedMember = memberCacheService.getMemberFromCache(memberId);
+            if (cachedMember != null) {
+                log.debug("캐시에서 현재 사용자 정보 조회 성공 - ID: {}", memberId);
+                return cachedMember;
+            }
         }
-
         Member member = memberDomainService.findById(memberId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEMBER_NOT_FOUND));
-
         MemberResponse response = MemberResponse.from(member);
-
-        memberCacheService.cacheMemberResponse(response);
-
+        if (memberCacheService != null) {
+            memberCacheService.cacheMemberResponse(response);
+        }
         log.debug("DB에서 현재 사용자 정보 조회 후 캐시 저장 - ID: {}", memberId);
-
         return response;
     }
 
     public MemberResponse getMemberProfile(Long memberId) {
         log.debug("회원 프로필 조회 요청 - ID: {}", memberId);
-
-        MemberResponse cachedProfile = memberCacheService.getMemberProfileFromCache(memberId);
-        if (cachedProfile != null) {
-            log.debug("캐시에서 회원 프로필 조회 성공 - ID: {}", memberId);
-            return cachedProfile;
+        if (memberCacheService != null) {
+            MemberResponse cachedProfile = memberCacheService.getMemberProfileFromCache(memberId);
+            if (cachedProfile != null) {
+                log.debug("캐시에서 회원 프로필 조회 성공 - ID: {}", memberId);
+                return cachedProfile;
+            }
         }
-
         Member member = memberDomainService.findById(memberId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEMBER_NOT_FOUND));
-
         MemberResponse response = MemberResponse.from(member);
-
-        memberCacheService.cacheMemberProfile(response);
-
+        if (memberCacheService != null) {
+            memberCacheService.cacheMemberProfile(response);
+        }
         log.debug("DB에서 회원 프로필 조회 후 캐시 저장 - ID: {}", memberId);
-
         return response;
     }
 
     public MemberResponse getMemberInfo(Long memberId) {
         log.debug("관리자 회원 정보 조회 요청 - ID: {}", memberId);
-
-        MemberResponse cachedMember = memberCacheService.getMemberFromCache(memberId);
-        if (cachedMember != null) {
-            log.debug("캐시에서 관리자 회원 정보 조회 성공 - ID: {}", memberId);
-            return cachedMember;
+        if (memberCacheService != null) {
+            MemberResponse cachedMember = memberCacheService.getMemberFromCache(memberId);
+            if (cachedMember != null) {
+                log.debug("캐시에서 관리자 회원 정보 조회 성공 - ID: {}", memberId);
+                return cachedMember;
+            }
         }
-
         Member member = memberDomainService.findById(memberId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.MEMBER_NOT_FOUND));
-
         MemberResponse response = MemberResponse.from(member);
-
-        memberCacheService.cacheMemberResponse(response);
-
+        if (memberCacheService != null) {
+            memberCacheService.cacheMemberResponse(response);
+        }
         log.debug("DB에서 관리자 회원 정보 조회 후 캐시 저장 - ID: {}", memberId);
-
         return response;
     }
 
@@ -173,13 +162,11 @@ public class MemberApplicationService {
     public MemberResponse approveMember(Long memberId, Long approverId) {
         Member member = memberDomainService.approveMember(memberId, approverId);
         MemberResponse response = MemberResponse.from(member);
-
-        memberCacheService.evictMemberCache(memberId);
-
-        memberCacheService.cacheMemberResponse(response);
-
+        if (memberCacheService != null) {
+            memberCacheService.evictMemberCache(memberId);
+            memberCacheService.cacheMemberResponse(response);
+        }
         log.info("회원 승인 완료 - ID: {}, approver: {}", memberId, approverId);
-
         return response;
     }
 
@@ -187,13 +174,11 @@ public class MemberApplicationService {
     public MemberResponse rejectMember(Long memberId) {
         Member member = memberDomainService.rejectMember(memberId);
         MemberResponse response = MemberResponse.from(member);
-
-        memberCacheService.evictMemberCache(memberId);
-
-        memberCacheService.cacheMemberResponse(response);
-
+        if (memberCacheService != null) {
+            memberCacheService.evictMemberCache(memberId);
+            memberCacheService.cacheMemberResponse(response);
+        }
         log.info("회원 거부 완료 - ID: {}", memberId);
-
         return response;
     }
 
@@ -201,13 +186,11 @@ public class MemberApplicationService {
     public MemberResponse suspendMember(Long memberId) {
         Member member = memberDomainService.suspendMember(memberId);
         MemberResponse response = MemberResponse.from(member);
-
-        memberCacheService.evictMemberCache(memberId);
-
-        memberCacheService.cacheMemberResponse(response);
-
+        if (memberCacheService != null) {
+            memberCacheService.evictMemberCache(memberId);
+            memberCacheService.cacheMemberResponse(response);
+        }
         log.warn("회원 정지 완료 - ID: {}", memberId);
-
         return response;
     }
 
@@ -218,17 +201,13 @@ public class MemberApplicationService {
                 request.getNickname(),
                 request.getEmail()
         );
-
         MemberResponse response = MemberResponse.from(member);
-
-        memberCacheService.evictMemberCache(memberId);
-
-        memberCacheService.cacheMemberResponse(response);
-        memberCacheService.cacheMemberProfile(response);
-
-        log.info("회원 프로필 업데이트 완료 - ID: {}, nickname: {}",
-                memberId, request.getNickname());
-
+        if (memberCacheService != null) {
+            memberCacheService.evictMemberCache(memberId);
+            memberCacheService.cacheMemberResponse(response);
+            memberCacheService.cacheMemberProfile(response);
+        }
+        log.info("회원 프로필 업데이트 완료 - ID: {}, nickname: {}", memberId, request.getNickname());
         return response;
     }
 
@@ -236,24 +215,24 @@ public class MemberApplicationService {
     public MemberResponse changeRole(Long memberId, Role role) {
         Member member = memberDomainService.changeRole(memberId, role);
         MemberResponse response = MemberResponse.from(member);
-
-        memberCacheService.evictMemberCache(memberId);
-
-        memberCacheService.cacheMemberResponse(response);
-
+        if (memberCacheService != null) {
+            memberCacheService.evictMemberCache(memberId);
+            memberCacheService.cacheMemberResponse(response);
+        }
         log.info("회원 권한 변경 완료 - ID: {}, role: {}", memberId, role);
-
         return response;
     }
 
     public MemberCacheService.CacheStatistics getCacheStatistics() {
-        return memberCacheService.getCacheStatistics();
+        return memberCacheService != null ? memberCacheService.getCacheStatistics() : null;
     }
 
     @Transactional
     public void evictAllMemberCache() {
-        memberCacheService.evictAllMemberCache();
-        log.warn("관리자에 의한 전체 회원 캐시 무효화 실행");
+        if (memberCacheService != null) {
+            memberCacheService.evictAllMemberCache();
+            log.warn("관리자에 의한 전체 회원 캐시 무효화 실행");
+        }
     }
 
     private void validateMemberStatus(Member member) {
